requirements.txt
numpy
dearpygui
mne-realtime
antropy
biotuner @ git+https://github.com/AntoineBellemare/biotuner.git
pyserial
oscpy
sounddevice
openai==1.3.6
scikit-learn
opencv-python==4.8.0.74 # prevent mediapipe TypeError: 'numpy._DTypeMeta' object is not subscriptable
pyzmq
webcolors
pylsl
mediapipe
librosa==0.9.2
panns_inference
networkx
deeptime
py-feat
geopy
edgeofpy @ git+https://github.com/jnobyrne/edgeofpy.git